+++ amps-maven-plugin/src/main/java/com/atlassian/maven/plugins/amps/util/minifier/GoogleClosureOptionsHandler.java
@@ -3,34 +3,23 @@ package com.atlassian.maven.plugins.amps.util.minifier;
 import com.google.javascript.jscomp.CompilerOptions;
 import org.apache.maven.plugin.logging.Log;
 
-public class GoogleClosureOptionsHandler
-{
+import javax.annotation.Nonnull;
 
-    private CompilerOptions compilerOptions;
-    private Log log;
+/**
+ * @deprecated This will be removed in AMPS 9. Use {@link com.atlassian.maven.plugins.amps.minifier.strategies.googleclosure.GoogleClosureOptionsHandler} instead.
+ */
+public class GoogleClosureOptionsHandler {
+    private final com.atlassian.maven.plugins.amps.minifier.strategies.googleclosure.GoogleClosureOptionsHandler impl;
 
-    public GoogleClosureOptionsHandler(Log log)
-    {
-        this.compilerOptions = new CompilerOptions();
-        this.log = log;
+    public GoogleClosureOptionsHandler(Log log) {
+        impl = new com.atlassian.maven.plugins.amps.minifier.strategies.googleclosure.GoogleClosureOptionsHandler(log);
     }
 
-    public void setOption(String optionName, String value) {
-        if (optionName.equals("languageIn")) {
-            setLanguageIn(value);
-        } else {
-            log.warn(optionName + " is not configurable.  Ignoring this option.");
-        }
+    public void setOption(@Nonnull final String optionName, final String value) {
+        impl.setOption(optionName, value);
     }
 
-    public CompilerOptions getCompilerOptions()
-    {
-        return this.compilerOptions;
-    }
-
-
-    private void setLanguageIn(String value)
-    {
-        compilerOptions.setLanguageIn(CompilerOptions.LanguageMode.valueOf(value));
+    public CompilerOptions getCompilerOptions() {
+        return impl.getCompilerOptions();
     }
 }