+++ amps-product-plugins/bamboo-maven-plugin/src/main/java/com/atlassian/maven/plugins/bamboo/BambooValidateBannedDependenciesMojo.java
@@ -1,0 +1,10 @@
+package com.atlassian.maven.plugins.bamboo;
+
+import com.atlassian.maven.plugins.amps.ValidateBannedDependenciesMojo;
+import org.apache.maven.plugins.annotations.Mojo;
+
+@Mojo(name = "validate-banned-dependencies")
+public class BambooValidateBannedDependenciesMojo extends ValidateBannedDependenciesMojo
+{
+
+}